* WhiteBoard:
  白板sdk
  删除[self.liveManager deleteMsg:sYSSignalVideoWhiteboard toID:YSRoomPubMsgTellAll data:nil completion:nil];
  no message
  单课件 白板视频标注信令
  全屏时的老师视频窗口
  删除多余代码
  优化登录时的选择框
  白板H文件

+++ YSLib/YSWhiteBoard/YSWhiteBoard.framework/Headers/YSDrawView.h
@@ -53,7 +53,7 @@ NS_ASSUME_NONNULL_BEGIN
  */
 - (void)setDrawType:(YSDrawType)drawType
            hexColor:(NSString *)hexColor
-           progress:(float)progress;
+           progress:(CGFloat)progress;
